Platform/requirements.txt
addict==2.4.0
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
filelock==3.16.1
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.2
Pillow==10.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3

numpy==2.1.2
pandas==2.2.3
scipy==1.14.1
sympy==1.13.3
mpmath==1.3.0

scikit-learn==1.5.2
scikit-image==0.24.0
matplotlib==3.9.2

torch==2.0.1
torchaudio==2.0.1
torchvision==0.15.2
transformers==4.45.2
diffusers==0.30.3
efficientnet_pytorch==0.7.1
huggingface-hub==0.25.2

opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
albumentations==1.4.21
grad-cam==1.5.4
retinaface-pytorch==0.0.7
ttach==0.0.3

datasets==2.13.0
modelscope==1.9.5
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
oss2==2.19.1

Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
streamlit==1.40.1

tqdm==4.66.5
networkx==3.3 